[xj:workflow:environment] default.env file
Environment (System) Properties

The default environment variables are in the file **/default.env**
which is copied to a new file **/.env** on project setup. Developers
modify their local .env file with private keys and configuration.

The **default.env** file is kept up-to-date with all
environment variables expected by this script.