Arrangement test #9 (bass movement)
https://www.pivotaltracker.com/story/show/181975107

We're hearing a problem with bass in slaps, where we aren't hearing enough movement in the choices of note from the voicings. We think that inversion seeking is prioritizing keeping all the notes as close together as possible for the segment. Instead, we'd like to preserve the motion inherent in the change from chord root to root.