Arrangement test #9 (bass movement) (#138)
https://www.pivotaltracker.com/story/show/181975107

We're hearing a problem with bass in slaps, where we aren't hearing enough movement in the choices of note from the voicings. We think that inversion seeking is prioritizing keeping all the notes as close together as possible for the segment. Instead, we'd like to preserve the motion inherent in the change from chord root to root.

**It turns out, we needed take semitone shift into account before computing octave shift! This is why Mark was hearing XJ adhering to the DP event instead of the target note.**