Ship irregular tempos https://www.pivotaltracker.com/story/show/182371790
Listen to Slaps in the latest app released for internal testing on app stores. It sounds like our HLS shipments system is unable to process irregular tempos. Assuming that the stream sounds fine in the template preview, while it is experiencing audio jumps/breaks/skips in the player app.

S O L V E D
-----------
Fully a math mistake I had made relating to calculating the position of those segment OGGs, that was fine when the positions were all even seconds (which happens at 120bpm with even numbers of beat lengths on everything) but blew up the moment that the tempos weren't multiples of 60