Run fabrication in background on exactly two threads, to avoid complexity
We're going to try using 2 threads, one for Control/Ship and one for Craft/Dub

Some mysterious hangs we're seeing are difficult to debug in the multithreaded paradigm. If it's not necessary, it would be great to avoid it completely.

However, it's important to thoroughly test this and make sure we don't introduce performance problems, e.g. gaps in music playback.

https://www.pivotaltracker.com/story/show/186696430